imp(template): fix func

+++ internal/repository/db/task/impl/delete.go
@@ -5,7 +5,7 @@ import (
 	"context"
 )
 
-func (r *repository) DeleteTask(id int) (err error) {
+func (r *repository) DeleteTask(id int64) (err error) {
 	_, err = r.Db.ExecContext(context.Background(), constant.DeleteTaskQuery, id)
 
 	return
